Fix circular dependencies in `scm-extra` package
